Ignore very recent archive files by default.
Added abilty to parse the hours and minutes fro archive filenames and default
the max timestamp to 10 minutes ago. This mitigates the race condition for job
begin and end archives where the archive indexer would try to read them after
they are created but before the data was flushed to the file by the compute
node.